+++ examples/node_sell_eth_infura.js
@@ -3,26 +3,25 @@
 const HDWalletProvider = require('truffle-hdwallet-provider')
 const sw = require('starkware_crypto')
 const Web3 = require('web3')
-
 const DVF = require('../src/dvf')
-const envVars = require('./helpers/loadFromEnvOrConfig')()
-const getPriceFromOrderBook = require('./helpers/getPriceFromOrderBook')
 
-const ethPrivKey = envVars.ETH_PRIVATE_KEY
-// NOTE: you can also generate a new key using:`
-// const starkPrivKey = dvf.stark.createPrivateKey()
-const starkPrivKey = ethPrivKey
-const infuraURL = `https://ropsten.infura.io/v3/${envVars.INFURA_PROJECT_ID}`
 
-const provider = new HDWalletProvider(ethPrivKey, infuraURL)
+
+const privateKey = '8F085...' // Account's private key
+const infuraKey = '9e28b...'  // Your Infura API KEY
+const infuraURL = 'https://mainnet.infura.io/v3/' + infuraKey
+
+const starkPrivKey = privateKey
+
+
+const provider = new HDWalletProvider(privateKey, infuraURL)
 const web3 = new Web3(provider)
-provider.engine.stop()
 
 const dvfConfig = {
   // Using staging API.
-  api: 'https://api.stg.deversifi.com'
-  // Add more variables to override default values
+  api:  'https://api.stg.deversifi.com'
 }
+
 
 ;(async () => {
   const dvf = await DVF(web3, dvfConfig)
@@ -30,14 +29,13 @@ const dvfConfig = {
   // Submit an order to sell 0.3 Eth for 200 USDT per 1 Eth
   const symbol = 'ETH:USDT'
   const amount = -0.3
+  const price = 200
   const validFor = '0'
   const feeRate = ''
 
-  // Gets the price from the order book api and cuts 5% to make sure the order will be settled
-  const orderBookPrice = await getPriceFromOrderBook();
-  const price = orderBookPrice - orderBookPrice * 0.05;
 
   const submitOrderResponse = await dvf.submitOrder({
+
     symbol,
     amount,
     price,
@@ -51,9 +49,14 @@ const dvfConfig = {
 
   console.log('submitOrder response ->', submitOrderResponse)
 
+
 })()
+// Stop provider to allow process to exit.
+.then(() => {
+  console.log('Stopping provider...')
+  provider.engine.stop()
+})
 .catch(error => {
   console.error(error)
   process.exit(1)
 })
-