+++ src/api/test/fixtures/getConf.js
@@ -45,13 +45,26 @@ module.exports = () => {
           '0x21ef21d6b234cd669edd702dd3d1d017be888337010b950ae3679eb4194b4bc',
         tokenAddress: '0x40d8978500bf68324a51533cd6a21e3e59be324a'
       }
-    },
-    spareStarkVaultId: 2090569095
+    }
   }
 
   nock('https://api.stg.deversifi.com')
     .post('/v1/trading/r/getConf', {})
     .reply(200, getConfResponse)
+
+
+    const mockGasResponse = {
+      cheap: 700000000,
+      average: 600000000,
+      fast: 500000000
+    }
+    
+    
+    nock('https://api.stg.deversifi.com')
+    .post('/v1/trading/r/getGasPrice', body => {
+      return true
+    })
+    .reply(200, mockGasResponse)
 
   const getUserConfResponse = {
     DVF: {
@@ -100,8 +113,7 @@ module.exports = () => {
       }
     },
     isRegistered: true,
-    ethAddress: '0xf858c2f2ac6b96df8c801bce90a3124a52d1915a',
-    recommendedGasPrice: 32500000000
+    ethAddress: '0xf858c2f2ac6b96df8c801bce90a3124a52d1915a'
   }
 
   nock('https://api.stg.deversifi.com')