+++ src/api/test/fixtures/mockGasPrice.js
@@ -59,4 +59,17 @@ module.exports = () => {
     .get('/json/ethgasAPI.json')
     .query(true)
     .reply(200, ethGasStationResponse)
+
+  const mockGasResponse = {
+    cheap: 700000000,
+    average: 600000000,
+    fast: 500000000
+  }
+  
+  nock('https://api.stg.deversifi.com')
+    .post('/v1/trading/r/getGasPrice', body => {
+      return true
+    })
+    .reply(200, mockGasResponse)
+
 }